Break out store cursors from default cursors
