+++ community/kernel-api/src/main/java/org/neo4j/internal/kernel/api/RelationshipDataAccessor.java
@@ -29,8 +29,6 @@ public interface RelationshipDataAccessor
 
     int type();
 
-    boolean hasProperties();
-
     void source( NodeCursor cursor );
 
     void target( NodeCursor cursor );
