+++ community/kernel/src/main/java/org/neo4j/kernel/impl/newapi/StoreRelationshipCursor.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2002-2018 "Neo4j,"
- * Neo4j Sweden AB [http://neo4j.com]
+ * Copyright (c) 2002-2018 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
  *
  * This file is part of Neo4j.
  *
@@ -21,118 +21,66 @@ package org.neo4j.kernel.impl.newapi;
 
 import org.neo4j.internal.kernel.api.NodeCursor;
 import org.neo4j.internal.kernel.api.PropertyCursor;
-import org.neo4j.internal.kernel.api.RelationshipDataAccessor;
 import org.neo4j.kernel.impl.api.RelationshipVisitor;
 import org.neo4j.kernel.impl.store.record.RelationshipRecord;
 
-abstract class RelationshipCursor extends RelationshipRecord implements RelationshipDataAccessor, RelationshipVisitor<RuntimeException>
+abstract class StoreRelationshipCursor extends RelationshipRecord implements RelationshipVisitor<RuntimeException>
 {
-    private boolean hasChanges;
-    private boolean checkHasChanges;
-    private PropertyCursor propertyCursor;
-    final DefaultCursors pool;
     Read read;
 
-    RelationshipCursor( DefaultCursors pool )
+    StoreRelationshipCursor()
     {
         super( NO_ID );
-        this.pool = pool;
     }
 
     protected void init( Read read )
     {
         this.read = read;
-        this.checkHasChanges = true;
     }
 
-    @Override
     public long relationshipReference()
     {
         return getId();
     }
 
-    @Override
     public int type()
     {
         return getType();
     }
 
-    @Override
     public boolean hasProperties()
     {
-        if ( read.hasTxStateWithChanges() )
-        {
-            PropertyCursor cursor = propertyCursor();
-            properties( cursor );
-            return cursor.next();
-        }
-        else
-        {
-            return nextProp != DefaultPropertyCursor.NO_ID;
-        }
+        return nextProp != NO_ID;
     }
 
-    @Override
     public void source( NodeCursor cursor )
     {
         read.singleNode( sourceNodeReference(), cursor );
     }
 
-    @Override
     public void target( NodeCursor cursor )
     {
         read.singleNode( targetNodeReference(), cursor );
     }
 
-    @Override
     public void properties( PropertyCursor cursor )
     {
         read.relationshipProperties( relationshipReference(), propertiesReference(), cursor );
     }
 
-    @Override
     public long sourceNodeReference()
     {
         return getFirstNode();
     }
 
-    @Override
     public long targetNodeReference()
     {
         return getSecondNode();
     }
 
-    @Override
     public long propertiesReference()
     {
         return getNextProp();
-    }
-
-    protected abstract void collectAddedTxStateSnapshot();
-
-    /**
-     * RelationshipCursor should only see changes that are there from the beginning
-     * otherwise it will not be stable.
-     */
-    protected boolean hasChanges()
-    {
-        if ( checkHasChanges )
-        {
-            hasChanges = read.hasTxStateWithChanges();
-            if ( hasChanges )
-            {
-                collectAddedTxStateSnapshot();
-            }
-            checkHasChanges = false;
-        }
-
-        return hasChanges;
-    }
-
-    // Load transaction state using RelationshipVisitor
-    void loadFromTxState( long reference )
-    {
-        read.txState().relationshipVisit( reference, this );
     }
 
     // used to visit transaction state
@@ -141,23 +89,5 @@ abstract class RelationshipCursor extends RelationshipRecord implements Relation
     {
         setId( relationshipId );
         initialize( true, NO_ID, startNodeId, endNodeId, typeId, NO_ID, NO_ID, NO_ID, NO_ID, false, false );
-    }
-
-    private PropertyCursor propertyCursor()
-    {
-        if ( propertyCursor == null )
-        {
-            propertyCursor = pool.allocatePropertyCursor();
-        }
-        return propertyCursor;
-    }
-
-    void close()
-    {
-        if ( propertyCursor != null )
-        {
-            propertyCursor.close();
-            propertyCursor = null;
-        }
     }
 }
